+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 - `kivik deploy local` deploys to a running Kivik instance, unless a deployment with key `local` exists in the RC file.
 - `kivik start` does not deploy by default.
+- RC file deployments can set a `fixtures` flag to determine if deployments should also deploy fixtures. This is false by default.
+- `kivik deploy` checks this flag (as well as the `local.fixtures` for `kivik deploy local`) when deploying.
 
 ## [2.0.0-alpha.7] - 2021-04-08